% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Function which adds two numbers x and y}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{First number which will be added}

\item{y}{Optional: Second number which will be added
default value is 10}
}
\value{
string which states who is amazing
}
\description{
Function which adds two numbers x and y
}
\examples{
my_add(1, 2)       # output: 3
my_add(1)          # output: 11
my_add(NA)         # output: NA
my_add(10, NA)     # output: NA
try(my_add("10", "20")) # output: error message saying "One of your inputs contains a string value"
}
